requirements.txt
fschat
fastapi==0.99.1
pydantic==1.10.7
sentence-transformers==2.2.2
uvicorn==0.23.1
tiktoken==0.4.0
numpy==1.24.4
scipy==1.10.1
scikit-learn==1.3.0
torchvision
torchaudio